Adjust the mutation multiplier over time
Linear interpolation that effects the mutation multiplier of each new generation. At the first new generation, the mutation multiplier is 2,. By the last generation, the mutation multiplier is 0.25. This means that the mutations are moving by a lot in the early generations, but as winning strategies emerge over time, the mutations are more subtle.