Bugfix - Game Crashes if only 1 player in game
The issue comes from some territories not having any armies on them because there wasn't enough armies to place them at the start of the game. The game winning conditions are checked at the end of each turn, so if there is a single player, they will not win until the end of their first turn. If that player tries to attack a territory with 0 armies on it, then the game will through an exception.